+++ packages/pwa-kit-dev/src/configs/webpack/config.js
@@ -29,7 +29,7 @@ import {CLIENT, SERVER, CLIENT_OPTIONAL, SSR, REQUEST_PROCESSOR} from './config-
 
 // Utilities
 import {getConfig} from '@salesforce/pwa-kit-runtime/utils/ssr-config'
-import {buildAliases, nameRegex} from '../../utils/extensibility-utils'
+import {buildAliases, getExtensionNames, nameRegex} from '../../utils/extensibility-utils'
 
 const projectDir = process.cwd()
 const pkg = fse.readJsonSync(resolve(projectDir, 'package.json'))
@@ -511,7 +511,7 @@ const requestProcessor =
 // Don't mistake this with the concept of extensions for Webpack.
 const extensions =
     mode === 'production'
-        ? (appConfig?.extensions || [])
+        ? (getExtensionNames(appConfig?.extensions) || [])
               .map((extension) => {
                   const setupServerFilePathBase = `${projectDir}/node_modules/${extension}/src/setup-server`
                   const foundType = ['ts', 'js'].find((type) =>